requirements.txt
streamlit
pandas
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib